requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.3.3
autopep8==1.5
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
comtypes==1.1.7
distlib==	0.3.1
dj-database-url==0.5.0
Django==3.1
django-crispy-forms==1.9.2
django-heroku==0.3.1
django-sequences==2.5
django-session-timeout==0.1.0
docopt==0.6.2
filelock==3.0.12
glob3==0.0.1
goto-statement==1.2
gunicorn==20.0.4
heroku==0.1.4
idna==2.10
image==1.5.32
imageio-ffmpeg==0.4.2
isort==4.3.21
Js2Py==0.70
keyboard==0.13.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
mhyt==0.7
mysql-connector-python==8.0.21
packaging==20.4
path==14.0.1
path.py==7.7.1
Pillow==7.2.0
pipwin==0.5.0
playsound==1.2.2
protobuf==3.12.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pyjsparser==2.7.1
pylint==2.4.4
pyparsing==2.4.7
PyPrind==2.11.2
PyQRCode==1.2.1
pySmartDL==1.3.3
python-dateutil==1.5
qrcode==6.1
requests==2.24.0
six==1.14.0
soupsieve==2.0.1
sqlparse==0.3.1
typing-extensions==3.7.4.2
tzlocal==2.1
urllib3==1.25.9
virtualenv==20.0.30
virtualenvwrapper-win==1.2.6
whitenoise==5.2.0
wrapt==1.11.2